feat(frontend): Implement voting system components and dashboard enhancements
